// frontend/web/src/services/API/proserve-wb-packaging-api/models/Image.ts
export interface Image {
    /**
     * Image ID
     * @type {string}
     * @memberof Image
     */
    imageId: string;
    /**
     * The build version of the image created by the pipeline
     * @type {string}
     * @memberof Image
     */
    imageBuildVersion: string;
    /**
     * The Amazon Resource Name (ARN) of the image created by the pipeline
     * @type {string}
     * @memberof Image
     */
    imageBuildVersionArn: string;
    /**
     * Pipeline ID that created the image
     * @type {string}
     * @memberof Image
     */
    pipelineId: string;
    /**
     * Pipeline Name that created the image
     * @type {string}
     * @memberof Image
     */
    pipelineName: string;
    /**
     * Recipe ID that is used by the pipeline
     * @type {string}
     * @memberof Image
     */
    recipeId: string;
    /**
     * Recipe Name that is used by the pipeline
     * @type {string}
     * @memberof Image
     */
    recipeName: string;
    /**
     * Recipe version ID that used by the pipeline
     * @type {string}
     * @memberof Image
     */
    recipeVersionId: string;
    /**
     * Recipe version Name that used by the pipeline
     * @type {string}
     * @memberof Image
     */
    recipeVersionName: string;
    /**
     * Status of the image (CREATING/CREATED/FAILED/RETIRED)
     * @type {string}
     * @memberof Image
     */
    status: string;
    /**
     * Amazon Machine Image (AMI) ID of the created image
     * @type {string}
     * @memberof Image
     */
    imageUpstreamId?: string;
    /**
     * Creation date of the image
     * @type {string}
     * @memberof Image
     */
    createDate: string;
    /**
     * Last update date of the image
     * @type {string}
     * @memberof Image
     */
    lastUpdateDate: string;
}